--- lazy.ts
@@ -1,11 +1,0 @@
-export function lazy<T>(f: () => T): () => T {
-    let called = false;
-    let result: T;
-    return () => {
-        if (!called) {
-            result = f();
-            called = true;
-        }
-        return result;
-    };
-}
