+++ test/lazy-test.ts
@@ -1,15 +1,15 @@
 import "mocha";
 import * as chai from "chai";
-import { lazy } from "../lazy";
+import { lazy } from "../index";
 
 chai.should();
 
 describe("lazy.ts", function() {
     it("lazy()", () => {
         let counter = 0;
-        const lazyFunc = lazy(() => { ++counter; return counter; });
+        const lazyFunc = lazy.lazy(() => { ++counter; return counter; });
         lazyFunc().should.equal(1);
         lazyFunc().should.equal(1);
         counter.should.equal(1);
-    })
-})
+    });
+});
