+++ test/syncmem-test.ts
@@ -15,7 +15,7 @@ describe("syncmem.ts", function() {
             syncMem.get(input.disjointUnion(bag.one(5)))().should.deep.equal([123, 5]);
             syncMem.get(input.product(bag.one([1, 2, 3]).flatten(x => x), (a, b) => [a * b]))()
                 .should.deep.equal([123, 246, 369]);
-        })
+        });
         it("get()", () => {
             const syncMem = new SyncMem();
             const r = bag.one("Hello world!");
@@ -26,6 +26,6 @@ describe("syncmem.ts", function() {
             const x = syncMem.get(input);
             syncMem.set(input, () => ["abc", "def"]);
             x().should.deep.equal(["abc", "def"]);
-        })
-    })
-})
+        });
+    });
+});
